ETR-40 Refactor: 워터마크 api 로직 리팩토링
