+++ src/pages/Watermark/WatermarkSuccess.tsx
@@ -1,5 +1,5 @@
-import axios from "axios";
 import { useLocation } from "react-router-dom";
+import { getWatermarkImage } from "../../api/watermark_api";
 
 function WatermarkSuccess() {
 
@@ -8,29 +8,7 @@ function WatermarkSuccess() {
 
   const handleDownload = async () => {
     try {
-      // 파일명 추출
-      const urlParts = downloadUrl.split('/');
-      const originalFileName = urlParts[urlParts.length - 1];
-
-      // 확장자 분리
-      const dotIndex = originalFileName.lastIndexOf('.');
-      const name = originalFileName.substring(0, dotIndex);
-      const ext = originalFileName.substring(dotIndex);
-
-      // 새로운 파일명 생성
-      const downloadFileName = `${name}${ext}`;
-
-      const response = await axios.get(`http://127.0.0.1:5000${downloadUrl}`, {
-        responseType: 'blob',
-      });
-      const url = window.URL.createObjectURL(new Blob([response.data]));
-      const link = document.createElement('a');
-      link.href = url;
-      link.download = downloadFileName;
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-      window.URL.revokeObjectURL(url);
+      await getWatermarkImage(downloadUrl);
     } catch (error) {
       alert("다운로드 실패")!
     }
